+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-09-08
 ### Added
 - Support for Basic/Digest auth username and password to be set via CLI options
 - Per-thread bandwidth limit option
@@ -28,7 +30,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Moved to github version of `pbr` to fix https://github.com/a8m/pb/pull/48
 
-[Unreleased]: https://github.com/daveallie/bindrs/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/daveallie/bindrs/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/daveallie/bindrs/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/daveallie/bindrs/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/daveallie/bindrs/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/daveallie/bindrs/compare/v0.1.0...v0.2.0